Move active range to res.config
This is perhaps not ideal. We would like code to move towards residing
in the ert package. However, this detail concerns the config and so the
use case is that you would like to parse parts of the config file and so
it belongs in a "config/" module. Also, it is used in the GUI, the
workflows and the ensemble evaluator, and has no dependents. Therefore
it needs to reside high in the import hirarchy to avoid circular
imports.
