+++ tests/ert_tests/cli/test_model_factory.py
@@ -2,7 +2,6 @@ from argparse import Namespace
 
 from ert_utils import ErtTest
 
-from ert.ensemble_evaluator.activerange import ActiveRange
 from ert_shared.cli import model_factory
 from ert_shared.libres_facade import LibresFacade
 from ert_shared.models.ensemble_experiment import EnsembleExperiment
@@ -10,6 +9,7 @@ from ert_shared.models.ensemble_smoother import EnsembleSmoother
 from ert_shared.models.multiple_data_assimilation import MultipleDataAssimilation
 from ert_shared.models.iterated_ensemble_smoother import IteratedEnsembleSmoother
 from ert_shared.models.single_test_run import SingleTestRun
+from res.config.active_range import ActiveRange
 from res.test import ErtTestContext
